Fix Rubinius builds hopefully
I'm unsure of how to do this in gemspec. This article[1] only mentions
Gemfiles. Not sure how to restrict a dependency of a gem to a specific
Ruby engine or even if that's a good idea.

However, it seems that if someone is using this tool from CLI, they are
probably using it with an installed MRI. If they are using it as a 
library, they are either on MRI (which doesn't need this dep), an earlier
version of Rubinius (which doesn't need this dep), or will already have
needed to require `rubysl` in their project anyway. So, putting it in the
Gemfile should be acceptable, at least for getting the tests running.

If someone comes across this and finds it to be less than ideal, pull
requests for a better approach are more than welcome.

[1]: http://www.benjaminfleischer.com/2013/12/05/testing-rubinius-on-travis-ci/